Update partner-xamarin-notification-hubs-android-get-started.md
I might have misunderstood what you are saying but for me the path is  

    https://HUB-NAMESPACE.servicebus.windows.net/HUB-NAME  

However after a lot of trial and error, the only thing working was giving the hub's name (as HUB-NAME in the path). Also, in the latest azure interface, the hub's name is indicated to be the path (in the quick glance section) which is confusing, that's why I added the comment.